consolidate some utils files
