+++ src/Components/Steps/HouseholdMembers/components/HouseholdMemberForm.tsx
@@ -17,13 +17,12 @@ import '../styles/IncomeSection.css';
 import { useShouldRedirectToConfirmation } from '../../../QuestionComponents/questionHooks';
 import useStepForm from '../../stepForm';
 import { LocationState } from '../utils/types';
-import { sortFrequencyOptions, calculateAge } from '../utils/calculations';
+import { sortFrequencyOptions, calculateAge, createHouseholdMemberData, scrollToFirstError } from '../utils/helpers';
 import { useIncomeStreamManagement } from '../hooks/useIncomeStreamManagement';
 import { useHouseholdMembersNavigation } from '../hooks/useHouseholdMembersNavigation';
 import { useHouseholdMemberConfig } from '../hooks/useHouseholdMemberConfig';
 import { createHouseholdMemberSchema } from '../utils/schema';
 import { createDefaultValues } from '../utils/defaultValues';
-import { createHouseholdMemberData, scrollToFirstError } from '../utils/formSubmission';
 import HealthInsuranceSection from '../sections/HealthInsuranceSection';
 import ConditionsSection from '../sections/ConditionsSection';
 import IncomeSection from '../sections/IncomeSection';
