+++ src/Components/Steps/HouseholdMembers/sections/ConditionsSection.tsx
@@ -11,10 +11,10 @@ import { ConditionOptions } from '../utils/types';
 import '../styles/HouseholdMemberSections.css';
 
 interface ConditionsSectionProps {
-  errors: FieldErrors;
+  errors: FieldErrors<any>;
   conditions: Conditions;
-  setValue: UseFormSetValue<any>; // Keep as any for flexibility with dynamic field updates
-  clearErrors: UseFormClearErrors<any>; // Keep as any for flexibility
+  setValue: UseFormSetValue<any>;
+  clearErrors: UseFormClearErrors<any>;
   options: ConditionOptions;
   pageNumber: number;
 }
